+++ src/functions/project-lint-terraform.ts
@@ -6,30 +6,31 @@ import {
 import { TerraformService } from '../services/index.js';
 
 /**
- * Implements the {@link ProjectLint} function for Terraform projects, by running `terraform validate` and
- * `terraform fmt`.
- * The Terraform format check is also run on any additional directories listed in the project configuration.
+ * The list of file extensions that `terraform fmt` supports.
+ */
+const TERRAFORM_FILE_EXTENSIONS = ['.tf', '.tfvars', '.tftest.hcl'];
+
+/**
+ * Implements the {@link ProjectLint} function for Terraform projects, by running `terraform fmt`.
+ * The Terraform format check is also run on any external files listed in the project configuration.
  */
 export class ProjectLintForTerraform extends ProjectLint {
   async _call(context: WorkspaceContext): Promise<void> {
     const projectPath = context.getProjectPathOrThrow();
     const projectName = context.get('project.name');
     const terraformService = context.service(TerraformService);
-    const externalPaths = await context.getProjectExternalPaths({
-      onlyDirectories: true,
-    });
-    const targets = [projectPath, ...externalPaths];
+    const externalPaths = await context.getProjectExternalPaths();
+    const externalTerraformFiles = externalPaths.filter((p) =>
+      TERRAFORM_FILE_EXTENSIONS.some((ext) => p.endsWith(ext)),
+    );
+    const targets = [projectPath, ...externalTerraformFiles];
 
     try {
       context.logger.info(
-        `🚨 Validating Terraform code for project '${projectName}'.`,
-      );
-      await terraformService.validate({
-        logging: { stdout: null, stderr: 'info' },
-      });
-
-      context.logger.info(
         `🎨 Checking format of Terraform code for project '${projectName}'.`,
+      );
+      context.logger.debug(
+        `Targets for Terraform format: ${targets.map((t) => `'${t}'`).join(', ')}.`,
       );
       await terraformService.fmt({
         check: true,